When a materialized view is first enabled, the consumer may be behind the transactional point at which the view was created.  In order to compensate for this, and not write an 'early' UOW_ID into the table, we generate a signal using mview_signal and record the signal_id into the mview table instead of the most recently available UOW_ID at the time of creation, thus preventing duplication of rows intothe mview on its first refresh.
mview_refresh() now checks to see if signal_id is not null and refreshed_to_uow_id is also null.  If this is true, we try to update the HWM entries with the real appropriate UOW_ID values from mview_uow, and failing this exit out of the refresh function using flexviews.SIGNAL().