Fix some problems with materialized views with aggregation, but no GROUP BY columns.
Now, when these views are reduced to zero rows during delta application, a zero count row is inserted into the table.
This accurately reproduces the same expected 0 count when there are no rows in the table.  All other columns will be NULL.

This probably won't work 100% correctly with views that have more than one count(*) column.  One of the columns will get a zero value and the others will be null.  I'll fix that later.

The auto_increment PRIMARY KEY is now added properly to these views.

test cases added (but not in this checkin)