Fixes: PERCENTILE had an issue when applying deltas.  This is resolved.
Reduce the number of temporary tables and subqueries for increased performance.

Keys are added to GROUP and COUNT expressions for improved refresh performance.

Full table scans of the materialized view have been eliminated during delta application.

Dead code removal

Reduce number of temporary table operations

Fix an issue with the order of expressions in the JOIN clause when tables are removed from a materialized view and added back again.