Make a number of fixes to FlexCDC.
a) Make FlexCDC exit with a non-zero error code whenever it exists on error.  This was done via replacing die(message) with a function called die1().
b) Fixed an issue with the timestamp data where where the value 0000-00-00 00:00:00 was always stored for timestamp columns.  
c) the setup script exits by default if it detects existing tables, added the --force command to force it to drop the tables before creating them again