Add the sleep_increment and sleep_maximum parameters and their description to the config file.
These two variables control the amount of time that FlexCDC spends sleeping between invocations of mysqlbinlog.
In order to avoid putting excessive load on the master, FlexCDC progressively backs off when it detects no changes
from the master.  By default it will back off in increments of .25 seconds up to five seconds, at which point it
will no longer increase the wait time any more.  The downside of this setting is that it can make incremental
refresh slower because the refresh process may have to wait for FlexCDC to wake up and process the changes that it
recently committed.